requirements.txt
Flask
spotdl
ffmpeg
python-dotenv
google-cloud-storage